pyproject.toml
[project]
name = "test-backend-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
